requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-cors-headers==3.13.0
django-filter==22.1
django-import-export==3.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
PyJWT==2.6.0
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.6
validate-docbr==1.10.0
xlrd==2.0.1
xlwt==1.3.0
